refactor(bucket): Omit type from generated bucket names
To ensure uniqueness and not overlapping with other configs, monaco created
buckets get a generated identifier based on their coordinate, similar to
other types like settings or automations.
Other than those types theses IDs are human readable, used in other configs
or manually and they only ever exist on the bucket API for bucket types.

As such every single bucket config will contain the unecessary string _bucket_
in it's 'bucketName'. As the type is not needed to ensure uniqueness and looks
weird to users, it is simply omited.

As this feature is not publicly released yet, not backwards compatibility to
the previous 'full coordinate' version is required.