fix(deploy): Honor continue-on-error and dry-run in new name validation
Unlike the old per-config validation for config API type's name uniqueness, the new
pre-flight validation always returned it's errors directly, even for a dry-run
(in which we want to collect all errors we can find) or when the continue-on-error
flag was set.

Now the options are honored and the validation errors are collected into the full
EnvironmentDeploymentErrors.

To enable this useage EnvioronmentDeploymentErrors.Append now accepts a variadic error
slice, so it can be used with single errors as well as the slice of errors from deployment.