fix(generate): Omit configs with empty names when generating deletefiles
Previously it was possible to generate invalid delete files if config had an empty string
set as name but was a Config API type. As persisting delete entries uses the same object
for all types it will omit empty names and IDs, and thus produce an invalid entry consisting
only of the 'type' if a config had an empty 'name' defined.