feat: Introduce default soft memory limit
It was found that with default settings, garbage collection
runs very seldomly during a deploymen, leading to memory consumption
to slowly build up and usually not go down before execution actuall
finishes.

This is likely due to the default GC setting of allocating up to 100%
of the current heap size within each next GC cycle.
With our memory pattern of allocating a large amount of
memory when loading all configurations at the start of a start of a
deployment, then slowly increasing memory with each deployed configuration
(rendered templates, logs, errors, etc.), this results in the GC mostly not
running after configs have been loaded.

As this in not a very reasonable memory behavior, and we can't expect all
users to set their own memory limits, a default limit of 2 GiB is introduced.

While hardcoding this isn't great, for now its a much simpler solution than
implementing OS specific ways to query the available memory and then still use
a hardcoded percentage.
Users that need finegrained control still have it by setting GOMEMLIMIT.