test(e2e): Fix bucket assert retry logic
The previous implementation didn't actually do what we want,
it didn't retry on 404 statuscodes, but only on unlikely
network errors, and it would have retrid immediately instead
of after a delay.

As such our E2E tests were flaky as it might still happen
occaisonally that an active bucket is not immediately returned.