refactor(settings_client): Do not use pointers and don't log all possible matches when finding objects with matching constraints
Using pointers to signify whether a match was found or not makes data unecessarily escape to the heap.
In this case using a struct and bool to signify if match was found works just as well.

Additionally including all objectIDs in the error message when many are found, potentially
creates a large error string, which is not really useful for a user to parse and is
currently kept in memory until deployment has finished - needlessly inflating in-use memory.