fix(settings_client): Consider Scope when looking for existing objects by unique property
The code looking for matching Settings to update based on unique properties defined by their schema
previously didn't take into account which scope a Setting is applied in.

As far as the Settings framework is concerned any uniqueness constraints applies with a single scope.
That means two settings can define the same unique property as long as they aren't in the same scope.

Our code now reflects this and only considers objects in the same scope when trying to find a matching
existing object based on the values of unique properties.

As this issue was discovered for process-group settings, an E2E test is added that defines settings using
the same unique PG meId but in different scopes.