refactor: Make log file paths available via exported functions
Several pieces of code like support archive creation or tests, rely on log
files being created with specific filenames in a specific folder.
Instead of replacing the strings everywhere, exported functions are
added to log and trafficlog, that allow anyone access to the expected
filepaths of log files for the current execution time(anchor).

As we don't want anyone to be able to modify this path, functions are
added instead of simply making the paths available as variables.