test(e2e): Assert that deployment errors are presented to users as expected
As only a very small handful of tests broke while refactoring the handling of errors
and their logging in deploy, a test is added to at least give some level of coverage
when messing with how errors are presented to users in the future.

Like other "produces_user_errors" tests, this is added as an E2E test, to ensure the
actual CLI as a whole product presents the information we want, instead of asserting
this on a lower package unit test level. E.g. deploy might still log correctly if you
give it the right logger, but the CLI might not if something in the setup code breaks
- which a CLI E2E test will still cover.

This test tries to strike a balance of testing specific enough messages to be sure
errors are still presented with the necessarry information, while not asserting
exact strings - that will hopefully allow small changes without constantly breaking
this test.