test: Remove unreasonable test forcing log file creation failure
The actuall filesystem and afero Fs do not fail a Mkdir/MkdirAll if a directory
already exists. The testcase using a specially implemted afero.Fs to test an
error case that should never produce an error in the real world is nothing but
missleading. If a .logs/ folder exists, we just want our logs written to it, not
get an error.