refactor(delete): Don't collect errors and reuse logger in Settings delete
To optimize heap memory consumption, the Settings object deletion,
is modified to log errors immediately rather than collecting them into a slice.

Additionally, instead of building several loggers with the same fields,
the functions are modified to build a logger object as early as possible
and reuse it. This results in fewer cloning operations of the zap logger
core and fields - which previously happened 1-2 times per log line, for
each .With... operation.