refactor(purge): Simply function calls to delete all
A simple delete.All is added to replace the individual All<Type>
functions previously called by purge.
If there is a need to delete All configurations of a specific type
again in the future, the now internal per-type packages could
be exported to expose their individual Delete and All functions.