fix(e2e): Ensure dependent test resources can be cleaned uo
Some test resources with dependendencies between each other, need to be removed
in the correct order. Not removing configs in order, e.g. lead to
automation business-calendars never beeing removed, as the test configs
also create scheduling rules which depend on those calendars.

To ensure test configs are fully cleaned up, the e2e test util
now sorts the project it is cleaning up and deletes in reverse
order of dependencies.
So while a businesss-calendar referenced in a scheduling rule is
deployed first, it is now removed last, after the scheduling rule
is gone.