refactor: Log execution errors ourselves instead of via cobra
Cobra logs errors returned when excuting a command using fmt.Fprtinf,
which results in a final log line that does not follow our log formatting
and is missing from logfiles.

To address this, we silence Cobra's builtin logging and instead log
errors ourselves in the CLI entry-point method Run().

Logging here, does impact logging in E2E tests, which build commands
directly and call Execute. In these cases the error is return and to
be handled by tests anyway though, so missing it from the logs the
tests produce should not have large impact.