refactor(template): Introduce new File and InMemory Templates
The updated FileBasedTemplate loads file contents only when required,
instead of at creation, reducing the long term in-use memory footprint
 of configurations.

The new InMemoryTemplate replaces the Download specific template, as
well as some dedicated test implementations of templates.
It currently carries an optional filepath, which is used for writing
converted templates to the desired target file.