fix(e2e): Cleanly handle missing API clients on cleanup
When cleaning up configurations configs that where skipped, don't need to
be deleted.
This also ensures that configs that might be explicitly skipped for non-platform
environments, don't fail to delete as no platform API clients exist.

To be fully save, platform deletions include a nil check on the respective client