refactor(persistence): Remove logic to persist templates that are never persisted
FileBasedTemplate is only used when a config was loaded from a file (in deployments), and
these types of Templates will never be passed to the config writer.
Only download and conversion will ever persist their configurations, and both create
InMemoryTemplates.

To simplify the persistence code and don't make it look like we ever persist FileBasedTemplates,
the codepath is simply removed.
If an unexpected type of Template still get's passed to the Writer, an error is returned.
If we ever extend Templates and don't implement persistence logic, we'll get an error.

As the interface already includes an error return and we won't notice this on compile,
an error was chosen over a hard-exit by panicking.