fix(generate): Include environment overrides when generating delete file and allow specifiying environments
Previously environment overwritten names for classic config API types where not
included in generated deletefiles.
This is fixed and by a full deletefile with any overrides is created.

To allow users to create more granular per-environment delete files if they
so desire, a --environment flag is introduced, that behaves the same as on
other commands.