fix(generate): Append timestamp before file-ending if generated delete.yaml already exists
As the delete command validates if it was passed a yaml file, by checking the file
ending, the current handling of avoiding overwriting existing files by appending a
filestamp after the full path, means that such generated deletefiles can't be used directly.
To address this, the timestamp is instead inserted before the file-ending.