chore(client): Allow deactivating update of single non-unique name cfgs by name
Updating single non-unique name configs can cause issues if a project defines more than one config
with the same name - they will overwrite each other, and keep a single on the environment.

As there currently does not seem to be fix without detrimental side-effects, the feature flag is introduced
to turn it off until a generally better solution is available.