feat: Skip updatable field in bucket download
The updatable field will be set by the API. No user can se updatable: false, so it does not make sense that we persist this property.

To implement this, we need to look into the json payload, thus a good place is the convertObject function. While it is not the perfect place, using errors, we can abstract the error nicely.