chore: Split e2e tests to execute them in parallel
This changeset reinvents how the e2e tests are executed.
First, instead of sequential steps, jobs are now executed in parallel.
Next, instead of two checkouts each step, we're querying the correct sha first, store it in an env var, and then use it in subsequent jobs.
Last, with this restructuring, we can skip jobs instead of steps using if conditions. This has the unfortunate consequence that we can't use env-vars in the ifs, so we need to hardcode the base-repo and tag-name.