refactor(account): Simplify usage of account delete client and document exported methods
Before this a caller of the delete client needed to be aware of the client - which in monaco's current
setup and DT account permission model only ever works for one account - as well as the account UUID the
client is for. This makes it possible to use the client for an account it has no permissions for.
To simplify this, the Client is not responsible of known the account UUID it is deleting from.