refactor(account): Separate logging of account deletion status from API client
This cleans up the API Client interface to do one thing again: delete resources and return
errors if something unexpted happens. The connection of API client and UUID in one struct is
still retained (see commit message for 'refactor(account): Split account delete command code into sub functions')
for reasoning.

To ensure the delete logic can differntiate deletion success for a resource not even being present,
an internal sentinel error is reused.

Additionally logging now includes both the name of an account as defined in the manifest, as well as
the UUID.