refactor(account): Load account resources into in-memory struct
So we're not directly using the persistence representation, the account loader
is modified to work like the one for configs and now returns the in-memory
struct.

With this, the persistence types are moved to an internal package shared only
by loader and writer. This necessitates to copy the structs to the current E2E
utility which also loads and re-writes the files to create unique names per test-run.
Re-using the loader and writer instead would modify the file structure, as the writer
creates a file per-type, not retaining the structure of files we have loaded from.
Thus the current solution needs to know about the persistence types.