refactor(manifest): Unify manifest value types
Several values in a manifest can have a type of 'environment' or 'value'.
So for one definition for URLs and one for Secretes existed. In preparation
for allowing Account UUIDs to be defined as hardcoded value or via env var,
the types are unified into a single struct and expected values.

Loading still ensures, that Secrets are only ever defined as env vars.