feat(account): Print a warning if configs are loaded in the project, and fail when account resources and configs are mixed.
Up to now, when loading AM resources, users would get no feedback whether AM resources and configs are mixed. This change now prints a warning if this happens, but continues loading.

However, when both configs and am resources are mixed in the same file, we're now failing hard if this is detected, as it is absolutly not allowed.