fix(account): Ensure optional parts of configurations are excluded when writing
Certain parts of configuration are not required - especially groups allow for permissions and
policies defined on several levels, but none are required. Before this change, the writer
required groups to have account level configurations, which was wrong, and would additionally
write empty slices into the YAML. Both of these are resolved.