fix: Fix settings update with overlapping ids
When updating a settings config, we try to 'onboard' existing configs whenever the originObjectId is set.
However, it can happen that there are two objects, one with the expected externalId, and a different one with the expected originalObjectId.
In this case, we update the setting with the externalId and print a proper warning.