fix(account): Ignore delete files when loading account resources
As the account loader reads any yaml in a folder, it was not possible to load
a folder containing a delete file, as it was attempted to be unmarshalled into
an account file struct.

An exclusion of delete files is added similar to the one for config files,
and account files are no longer unmarshalled strictly, as we gain no clear
benefit from creating an error for any unknown field in code that loads everything
that has a yaml file type.