fix: Don't skip downloading modifiable configs
We were skipping all configs that had the flag property 'modifable: false' set. While this is correct in most cases, the 'modiablePaths' property can overwrite this for specific paths.
In this case `(modifiable == false && modifiablePaths not empty)` , we are now downloading the config and persisting it.
This behavior is valid for the following reasons:
1) Upload the config to the same environment:
1.1) Updating enabled -> valid use case, needs to be supported
1.2) Updating anything else -> API returns an error informing the user.
2) Upload to another environment:
2.1) Object is predefined and has the same ID everywhere: We update enabled correctly, everything else is reported by the API as invalid.
2.2) Object is not predefined: We are creating it, behavior depends on the config.