feat(manifest): Make envGroups and accounts optional
Instead of always requting environmentGroups and viewing accounts as
optional when loading a manifest, they are now both optional.
This means manifests that are only used for deploying account resources,
no longer need to define envirnment groups.

For convenience validation of manifests is still done in the loader,
allowing callers to optionally define whether they require envs or
accounts to be defined. With those options set, they do not be extended
with dedicated validations, but just react to errors of the loader function
as before.