fix: Extend the dependency resolution to not extract some substrings
The issue fixed is, that we sometimes have small IDs that are the substring of another id.

For example, the metric-id "calc:test" was found in the metric-id "calc:testcase", which lead to the incorrect substitution "{{ .parameter}}case".
From manual investigation, these prefixes/suffixes that are alphanumeric, _, and - are most often false positives, thus we include them in our search.

Additionally, this change has the impact that 'plain' strings won't be found. For example, if a to-be-searched string is only `1234`, the id can't be found, as there is an additional character needed in the front and back. In reality, this is not an issue as we always handle content in json documents that has at least quotes around it.