fix(download): Correctly limit downloads based on options
Previously download always downloaded classif config types as well as grail buckets.
In both cases it wrongly ignored options that would limit the download to other types
in general or specific APIs or schemas.

This is fixed for both types now, excluding them from download if other types are requested.