refactor(account): Remove unnecessary sanity check
This check verified whether resp.count == resp.Items.Count and failed if the condition was not met.
For a user, this check is meaningless, as they can do nothing if it fails, and not being able to use monaco in that case until it is fixed does not make sense.