ci: Use unique artifact names for test results
With the v4 release of the publish artifacts GH action, it is no
longer possible to upload additional files to the same artifact.

To rememdy this, each build step now uploads an artifact with it's
own unique name.
The result collection workflows should not impacted by this, as they
simply download all artifacts of the triggering job and unpack them.