fix: Hack integration tests together that 'name' is not replaced in e2e tests.
Our e2e test framework appends a suffix to all yaml values where the key starts with `name`. Since the API is not a generic name, it must not be changed. To counter that, without rewriting the whole test framework, a hint `#monaco-test:no-replace` is introduced to keep some lines as is, if wished.

This is not a nice change, but without changing the whole replacement it's the best I came up with.