refactor: Simplify type loading
This PR is the first step to simplify the type loading.
Instead of unmarshalling into an intermediate object, we unmarshal directly into the resulting object.
This skips the extra unmarshalling steps later and allows you to access the information directly.

This can be improved even further if the scope is completely unmarshalled during the config unmarshalling.