fix: Replace SubPath method by SubPath Property
The SubPath method broke the deployment as the former method checked whether the url contained the {SCOPE} sub path.
This leads (and led) to a problem when we depended in a later step on the subpath check, and at that time, the proper url without the {SCOPE} was already used instead.