fix: calculated-metrics-service should NOT remove the property `enitityId`.
`calculated-metrics-service`'s entityId points to a service, which should be extracted to a parameter.
For most other APIs (note: this check is only performed for classic APIs, not settings), the entityId points to itself, so it must be removed.
For some (synthetic-test), the entityId points to the steps, which are not handled by monaco (and they don't make sense to) so they are removed as well.