refactor: Drop manifest version back to 1.0
When account resources where originally introduced, we expected
the manifest extension to support them to be a breaking change.

In their current version the addition of accounts to the manifest
is not a breaking change, but they are just an optional field in
the manifest. Older versions targeting config deployments will already
produce an error we consider clear enough, mentioning that the accounts
field can not be unmarshalled, and do not offer an account deployment
command, so it is should be clear enough to users that they need a
newer version to work with account resources.