fix: Duplicate name check must take environment into account
The duplicate name check marks all configs that are non-unique configs and where the name is declared as duplicate. This has the intention, that we don't update the same object multiple times in the same source config set.

However, a bug was present that if the same config was defined for multiple environments, that the duplicate name flag was set, since the config existed multiple times with the same name, **but for different environments**.

This is now fixed by taking the target env into account during the check