fix(client): Don't attempt to build simple classic URL if source URL is not a default platform URL
Currently if a user wrongly configures a classic URL with an oAuth client, a simple classic
URL will be built for it (not replacing anything from the valid classic URL) and we'll continue
with the assumption that the client is now configured for a platform and classic URL.
Once we attempt to make platform API calls against the configured URL, these calls will
result in hard to decypher errors.

The addition of creating a classic URL without making an API call, has thus removed our
check and error messages catching if the given URL is not for a platform environment at
all.