fix: ensure web key user action's scope parameter is a reference parameter
This ensures that key user actions for web applications are in the same
independent component (subgraph to be deployed). We need to do this because we need
a way to deploy them sequentially with a delay later on, because the API does not like parallel requests. KUAWs with hard coded scopes would result in monaco not knowing the dependency between the KUAW and the referenced (hard-coded) Web app, hence they would end up in different independent graph components making it impossible to deploy sequentially with a delay.