ci: Only run IAM E2E tests if specifically triggered via label
As there is a known issue in the IAM API, these tests are not currently useful
to us.
To still be able to run them selectively, they are limited to unly run for
PRs if labeled with a specific new label.
This is to be removed when the API issues are resolved and we actually get
value from these tests again