feat: introduce MONACO_LOG_FILE environment variable
MONACO_LOG_FILE_ENABLED is treated to be set per default, hence monaco will continue writing to the .logs directory. If it is set to false, it will not write log files. This is overruled as soon as the --support-archive flag is used. Then we obviously need/want to write logs to files and collect them.