fix(download): Strip conversionGoal IDs from application-web JSON payload
When downloadng application-web configs that define conversion goals, we so far
retained the goal's IDs in the payload.
On re-upload this works, but when the downloaded application config is applied
to an empty environment the POST call fails if these IDs are sent.

Thus the download sanitizer is extended to include this nested ID field.
Note that it'd be nice to have a more general removal of nested 'id' fields
in the future - but this fix is in line with similar removed fields.