docs: Describe at which point full logging is available
As we allow users to control things like log versobity via a cmd line flag,
we set up our logger only just before running a cobra command - using a PreRun
method of the Command object. To make it clear what the earliest point for
full logs is, the comment on the initial temporary logger setup in main is
extended with a note pointing the reader the right way.