fix: avoid dots in extracted parameters
during download, we extract and convert IDs into parameters, we also do this for the "scope" parameter.
Depending on the settings object the scope could contain dots (".") which are then used as a key of the reference parameters
However, monaco fails to resolve these parameters as monaco assumes dots to be nested structures. Example:

- id: 1
  config:
    parameters:
      extractedIDs:
        type: value
        value:
          id_metric_postgres.session_time.count: metric-postgres.session_time.count
    template: 1.json
    skip: false
  type:
    settings:
      schema: builtin:metric.metadata
      schemaVersion: "1"
      scope:
        property: extractedIDs.id_metric_postgres.session_time.count
        type: reference

monaco fails to resolve the parameter "extractedIDs.id_metric_postgres.session_time.count" as it assumes the following structure:
parameters:
      extractedIDs:
        type: value
        value:
          id_metric_postgres:
            session_time:
              count: metric-postgres.session_time.count

This PR avoids running into this situation by simply making sure that parameter keys do not contain dots.