build: Use ko for building containers
ko is a simple fast image builder for Go, similar to jib for Java.
It simplifies building images without the need for docker to be installed
as it's fully Go based, and supports multi-plaform images as well as generating
SBOMs.

It is also the recommended tool for building containers with GoRelaser, so a
future switch of the release pipeline to using GoReleaser would simply mean
transfering the config to the semantically equal config for that tool.

see: https://ko.build/