build: Use explicit latest Go toolchain version for builds
Our currently used go directive omitting the patch version,
will result in .0 'or higher' being used at the discrection
of the build toolchain. In some tools this open defintion
leads to 1.22.0 being used (happens in ko container image
build), which is 6 patches behind and has vulnerabitlies.
We also can't be sure in the regular builds that the latest
patch is being used or not.

Thus an explicit fixed version is set via the toolchain directive
introcduced in Go 1.21. Fixing the compiled toolchain to the
currently latest 1.22.6 patch.
Our setup of dependabot should cover these version updates as
well and update with future patches automatically.