build: Install ko as part of docker-container Make target
Similar to other targets, this ensures the target can be used directly as
it installs all tools it needs automatically.
With this the dedicated install step from the release pipeline is removed
as well - as the builds for internal and public images run sequentally in
the same step/container anyway, the second will find the binary installed
and not need install it twice, so the extra 'install' call has no negative
impact.