build: Use extended ko allowing us to overwrite the binary base path
Ko currently places binaries in a 'ko-app/' directory by default, which
would produce a subtly different image to the one we've built with Docker
/BuildKit before. For CI systems like GitLab, where the entry-point of
the image needs to be overwritten with a shell, this would mean that an
image release using ko will break customer's builds, as they won't find
the 'monaco' command directly runnable (in /usr/local/bin) anymore.

As an open issue for ko exists to make the base app directory configurable,
a fork was created allowing this.
Until the proposed change is reviewed and approved by the ko maintainers,
we can use the forked version directly to ensure the change of build tool
is transparent to our customers.