build: Split installing ko into dedicated Make target
As the release build needs ko not only for building an image
but also to login to the container registry, we still need to
be able to install it seperately from the makte target for
building a container.

To keep the make target convenient for local use, and ensure
we only need to switch/update ko versions in one place, a new
make target is added for installing ko, and used in the release
pipeline to install it as an initial step.