fix: better max filename detection
This approach works also if due to deduplication of file names appending a suffix to the filename results in exceeding the limit again. Sanitizing files names and Sanitizing strings are separate now. Only when writing the template files, the file name deduplicate logic and filename length checks are done. In case there are still troubles with the default max length (=254) the user can override it using "MONACO_MAX_FILENAME_LEN" environment variable. Note that 2 places of the max fileName lenght are actually reserved for the duplication counter.