fix(dependency_resolution): Dashboard-Share-Settings can no longer be referenced from dashboards.
Dashboard share settings are a bit special. One dashboard always has exactly 1 dashbaord share setting connected to it. This share setting has the same ID as the dashboard itself. This is were the issues arise.

Dashboards can't depend on other dashboards, as we ignore those references. They are usually only included in the markdown of the tiles, not the dashboards themselves. Since the dashboard share settings have the same ID as the dashboards, what means that instead of ignoring dashboard->dashboard references, we get unwanted dashboard->dashboard-share-settings references. What is more, the references are not within the same config, but some random dashboards can suddenly depend on the dashboard-share-settings.
Since dashboard-share settings do depend on their parent dashboard, it can happen that we suddenly introduce unwanted cyclic dependencies.

This fix ignores dashboard-share-settings completely as nothing can actually reference them.