fix(Settings): App settings can now be created
App-settings work a bit different from builtin-settings. When creating app-settings, the originObjectId must not be set, while the 'normal' settings do allow that.

To fix that, the approach has been changed. We assume that we can't find the correct object that requires updating, and then search for it.
First, by the unique-constrains of the schema, then by the legacy-external-id, and finally by the external-id and origin-id.

Note that currently, the later methods potentially overwrite the previous methods. This can be fixed in a later refactoring.

Furthermore, this commit introduces multiple new tests that test the behavior and fixes existing ones (plus correct assert calls).