fix: Add `afterInsert` validator
Validating `afterInsert` parameters is only possible in some cases:
* The validation is only performed before deployment on static data, not during deployment on dynamic data. The API will return an error any way.
* The `afterInsert` parameter must be a reference parameter
* The scopes of both configs must be a simple value parameter.

Only if all those preconditions apply, we can do the proper check. Additional to checking whether the scope is the same, the validator also verifies that the schema is the same, as well as the reference actually exist.

Furthermore, with this change some other tests had to be fixed