fix: File parameter must provide references (#1760)
This PR restricts file parameters to only using properties defined via
`references`. Concretely this means that if you define

```
parameters:
  request:
    type: file
    path: request.json
    escape: false
  ua_string: Mozilla/5.0...
```

the template file parameter `request` cannot `ua_string`. To allow this,
this must be explicit in the `references`:

```
parameters:
  request:
    type: file
    path: request.json
    escape: false
    references:
    - ua_string
  ua_string: Mozilla/5.0...
```

This is actually how the file parameter was designed and documented, it
was just never enforced in the original implementation.