feat(settings): add download for ACL all-users (#1777)
#### What this PR does / Why we need it:
Introduces the download functionality for ACL enabled settings (behind a
feature flag)

#### Special notes for your reviewer:
In order to get the ACL property, the schema has to be fetched (not
included in the GetAll API call).

At the moment I'm fetching all needed schemas at once. Tell me if they
should not be in parallel