feat(settings): Require OAuth for Settings config with defined `permission` (#1782)
<!--  Thanks for sending a pull request! 
If this is your first time, please read our contributor guidelines:
https://github.com/Dynatrace/dynatrace-configuration-as-code/blob/main/CONTRIBUTING.md

Before submitting this PR, please make sure that:
1. Your code builds without any errors or warnings
2. Your code is covered by unit tests
3. Your branch is rebased on top of current main (`git pull --rebase
origin main`)
-->

#### What this PR does / Why we need it:
Validation that OAuth is defined for the usage of Permissions on the
`SettingsType`.

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
<!--
If no, just write "NONE" or leave this section empty.
If yes, state how the user is impacted by your changes.
-->