feat(settings): add deploy for ACL all-users (#1778)
#### What this PR does / Why we need it:
Adds support for ACL "all-users" in settings on deploy

#### Special notes for your reviewer:
I added some testing utils, which reduce the overhead for the testing
server a lot.
I also added an exception for 404 of the permission deploy. We can
discuss this one. Currently a delete is always one call. If we would add
a GET check before it would be 1 (not exists) to 2 (exists, GET +
DELETE) API calls