feat: Sanitize bucket names (#1815)
This PR adds sanitization of generated bucket names:
- Bucket names are still generated by concatenating project name with
the config id: `[projectName]_[configId]`
- Bucket names are sanitized by:
  - converting the name to lower case
- removing characters such that the result must match
`([a-z])([a-z0-9])([a-z0-9_-])+`
  - truncating to a maximum length of 100 characters.

By default this is enabled, however this can be disabled by setting the
feature flag `MONACO_SANITIZE_BUCKET_NAMES` to `false`.

Furthermore, it makes sure the generated bucket name is visible in logs
and errors by updating several error messages as well as updating the
version of the core library used.

---------

Co-authored-by: Klaus Strießnig <klaus.striessnig@dynatrace.com>