fix: don't print version warning for OAuth (#1819)
#### What this PR does / Why we need it:
If OAuth is in use, the cluster version check is not needed, as it will
always be supported.
It did not work anyway because a classic endpoint was used for platform.

The following warning log was printed if OAuth was used on download
```
Unable to determine server version "https://xyz.apps.dynatrace.com":  ...
```
now we aren't checking the version for OAuth anymore and the warning
message will not be printed.

There are two other places in the code where the version check happens,
but these two are using the classic client and not the platform one.

#### Special notes for your reviewer:
Tests are not added and would require some refactoring. We want to split
the command code in the future to not contain any logic, therefore
leaving this out atm.

#### Does this PR introduce a user-facing change?