refactor(deploy): cache per environment (#1831)
#### What this PR does / Why we need it:
Fills the cache per environment when it's deployed. So instead of
filling the cache with all environments we are filling it when we deploy
the config for one environment (only one env in cache at the time).


#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?
Yes, less memory consumption if several environments are used. Only one
environment will be cached at the time